// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/OutstandingDebtDocumentRequest.ts
export interface OutstandingDebtDocumentRequest {
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentRequest
     */
    loanProvider?: string;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentRequest
     */
    currency?: string;
    /**
     * 
     * @type {number}
     * @memberof OutstandingDebtDocumentRequest
     */
    remainingBalance?: number;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentRequest
     */
    repaymentType?: OutstandingDebtDocumentRequestRepaymentTypeEnum;
    /**
     * 
     * @type {Date}
     * @memberof OutstandingDebtDocumentRequest
     */
    finalRepayment?: Date;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentRequest
     */
    securityType?: OutstandingDebtDocumentRequestSecurityTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof OutstandingDebtDocumentRequest
     */
    revenueShare?: number;
    /**
     * 
     * @type {number}
     * @memberof OutstandingDebtDocumentRequest
     */
    maximumLimit?: number;
}


/**
 * @export
 */
export const OutstandingDebtDocumentRequestRepaymentTypeEnum = {
    Amortising: 'AMORTISING',
    Bullet: 'BULLET',
    RevolvingLineOfCredit: 'REVOLVING_LINE_OF_CREDIT',
    ConvertibleLoan: 'CONVERTIBLE_LOAN',
    RevenueBasedFinancingMca: 'REVENUE_BASED_FINANCING_MCA',
    CreditCardDebt: 'CREDIT_CARD_DEBT',
    Leases: 'LEASES'
} as const;
export type OutstandingDebtDocumentRequestRepaymentTypeEnum = typeof OutstandingDebtDocumentRequestRepaymentTypeEnum[keyof typeof OutstandingDebtDocumentRequestRepaymentTypeEnum];

/**
 * @export
 */
export const OutstandingDebtDocumentRequestSecurityTypeEnum = {
    Secured: 'SECURED',
    Unsecured: 'UNSECURED'
} as const;
export type OutstandingDebtDocumentRequestSecurityTypeEnum = typeof OutstandingDebtDocumentRequestSecurityTypeEnum[keyof typeof OutstandingDebtDocumentRequestSecurityTypeEnum];